% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityPlot.R
\name{densityPlot}
\alias{densityPlot}
\title{Generates a kernal density plot}
\usage{
densityPlot(samples = 50, sample_size = 100, mean = 0, sd = 5)
}
\arguments{
\item{samples}{Specify how many samples to run}

\item{sample_size}{the number of observations in each sample}

\item{mean}{population mean}

\item{sd}{population standard deviation}
}
